# PR: Implement Benchmark Module with Optimized Database Configuration