--- a/benchmarks/__init__.py
+++ b/benchmarks/__init__.py
@@ -1,0 +1,12 @@
+"""
+Benchmarks package for AgentFarm.
+
+This package provides a framework for creating, running, and analyzing benchmarks
+for the AgentFarm system.
+"""
+
+from benchmarks.base.benchmark import Benchmark
+from benchmarks.base.runner import BenchmarkRunner
+from benchmarks.base.results import BenchmarkResults
+
+__all__ = ["Benchmark", "BenchmarkRunner", "BenchmarkResults"]