--- a/benchmarks/base/__init__.py
+++ b/benchmarks/base/__init__.py
@@ -1,0 +1,3 @@
+"""
+Base classes for the benchmarks package.
+"""