--- a/benchmarks/examples/__init__.py
+++ b/benchmarks/examples/__init__.py
@@ -1,0 +1,3 @@
+"""
+Example scripts for using the benchmark framework.
+"""