--- a/benchmarks/implementations/__init__.py
+++ b/benchmarks/implementations/__init__.py
@@ -1,0 +1,7 @@
+"""
+Benchmark implementations.
+"""
+
+from benchmarks.implementations.memory_db_benchmark import MemoryDBBenchmark
+
+__all__ = ["MemoryDBBenchmark"]