This PR introduces a comprehensive benchmark module for AgentFarm that enables systematic performance testing, analysis, and configuration optimization. Based on benchmark results, we've implemented a recommended configuration for simulations requiring post-simulation analysis that provides a 33.6% performance improvement while maintaining full data persistence.

## Key Features

### 1. Benchmark Framework
- Created a modular, extensible benchmark framework with clear separation of concerns
- Implemented base classes for creating, running, and analyzing benchmarks
- Added utilities for visualization and statistical analysis of benchmark results
- Designed for easy extension with new benchmark types

### 2. Memory Database Benchmark
- Implemented a benchmark comparing disk-based and in-memory database performance
- Tested three configurations:
  - Disk-based database (baseline)
  - In-memory database without persistence (maximum performance)
  - In-memory database with persistence (balanced approach)
- Results show significant performance improvements:
  - In-memory without persistence: 44.6% faster than disk-based
  - In-memory with persistence: 33.6% faster than disk-based

### 3. Configuration Helpers
- Added utility functions to easily apply the recommended configuration
- Created a helper script to generate optimized configurations for simulations
- Implemented detailed documentation and examples

### 4. Example Scripts
- Added example scripts demonstrating how to use the benchmark framework
- Created a script for running simulations with the recommended configuration
- Included comprehensive documentation

## Directory Structure
```
benchmarks/
├── base/                  # Base classes for benchmarks
│   ├── benchmark.py       # Abstract base class for all benchmarks
│   ├── runner.py          # Class for running benchmarks
│   └── results.py         # Class for storing and analyzing results
├── implementations/       # Concrete benchmark implementations
│   └── memory_db_benchmark.py  # Memory database benchmark
├── results/               # Directory for storing benchmark results
├── utils/                 # Utility functions
│   ├── visualization.py   # Visualization utilities
│   ├── statistics.py      # Statistical analysis utilities
│   └── config_helper.py   # Configuration helper utilities
├── examples/              # Example scripts
│   └── run_with_recommended_config.py  # Example using recommended config
├── run_benchmarks.py      # Main script for running benchmarks
└── README.md              # Documentation
```

## Benchmark Results
```
Configuration: 100 steps, 30 agents, 3 iterations
Average disk database time: 5.99 seconds
Average in-memory database time (without persistence): 3.31 seconds
Average in-memory database time (with persistence): 3.98 seconds
Performance improvement (without persistence): 44.6%
Performance improvement (with persistence): 33.6%
```

## Recommended Configuration
Based on the benchmark results, we recommend using in-memory database with persistence for simulations requiring post-simulation analysis:

```python
config.use_in_memory_db = True
config.persist_in_memory_db = True
```

This configuration provides:
- **33.6% faster execution** than disk-based database
- **Full data persistence** for post-simulation analysis
- **Good balance** between performance and data durability

## Validation
We've validated the recommended configuration with a full simulation run:
- 30 agents (10 of each type)
- 100 simulation steps
- Completed in 2.42 seconds
- Successfully persisted 7,986 rows across 9 tables in 0.10 seconds

## Usage Examples

### Running a Simulation with Recommended Configuration
```bash
python -m benchmarks.examples.run_with_recommended_config
```

### Using the Configuration Helper in Code
```python
from benchmarks.utils.config_helper import configure_for_performance_with_persistence

# Apply recommended settings to an existing config
config = SimulationConfig()
config = configure_for_performance_with_persistence(config)
```

### Running Benchmarks
```bash
python -m benchmarks.run_benchmarks --benchmark memory_db --steps 200 --agents 50 --iterations 5
```

## Future Work
- Add more benchmark types (e.g., agent behavior, resource distribution)
- Implement automated benchmark runs for CI/CD pipeline
- Create a dashboard for visualizing benchmark results over time
- Explore additional optimization opportunities

## Testing
- Manually tested all benchmark functionality
- Verified results consistency across multiple runs
- Validated persistence functionality with post-simulation analysis

## Documentation
- Added comprehensive README with usage examples
- Included detailed docstrings for all classes and functions
- Created example scripts demonstrating key functionality